bernoulli, binomial, categorical: make CDFs càdlàg

+++ benches/distribution/binomial.rs
@@ -6,7 +6,7 @@ fn cdf(bencher: &mut Bencher) {
     let binom = Binomial::new(100_000, 0.845);
     let x = Independent(&binom, &mut random::default()).take(1000).collect::<Vec<_>>();
 
-    bencher.iter(|| black_box(x.iter().map(|&x| binom.cdf(x)).collect::<Vec<_>>()));
+    bencher.iter(|| black_box(x.iter().map(|&x| binom.cdf(x as f64)).collect::<Vec<_>>()));
 }
 
 #[bench]
